+++ lib/geo/geometry_collection.ex
@@ -20,7 +20,7 @@ defmodule Geo.GeometryCollection do
 
   def cast(%Geo.GeometryCollection{} = geometry_collection), do: {:ok, geometry_collection}
   def cast(geometry_collection) when is_map(geometry_collection), do: { :ok, Geo.JSON.decode(geometry_collection) }
-  def cast(geometry_collection) when is_binary(geometry_collection), do: { :ok, Geo.JSON.decode(geometry_collection) }
+  def cast(geometry_collection) when is_binary(geometry_collection), do: { :ok, Poison.decode!(geometry_collection) |> Geo.JSON.decode }
   def cast(_), do: :error
 
 end