+++ lib/geo/line_string.ex
@@ -20,7 +20,7 @@ defmodule Geo.LineString do
 
   def cast(%Geo.LineString{} = line_string), do: {:ok, line_string}
   def cast(line_string) when is_map(line_string), do: { :ok, Geo.JSON.decode(line_string) }
-  def cast(line_string) when is_binary(line_string), do: { :ok, Geo.JSON.decode(line_string) }
+  def cast(line_string) when is_binary(line_string), do: { :ok, Poison.decode!(line_string) |> Geo.JSON.decode }
   def cast(_), do: :error
 
 end