+++ lib/geo/multi_point.ex
@@ -20,7 +20,7 @@ defmodule Geo.MultiPoint do
 
   def cast(%Geo.MultiPoint{} = multi_point), do: {:ok, multi_point}
   def cast(multi_point) when is_map(multi_point), do: { :ok, Geo.JSON.decode(multi_point) }
-  def cast(multi_point) when is_binary(multi_point), do: { :ok, Geo.JSON.decode(multi_point) }
+  def cast(multi_point) when is_binary(multi_point), do: { :ok, Poison.decode!(multi_point) |> Geo.JSON.decode }
   def cast(_), do: :error
 
 end