+++ lib/geo/point.ex
@@ -20,7 +20,7 @@ defmodule Geo.Point do
 
   def cast(%Geo.Point{} = point), do: {:ok, point}
   def cast(point) when is_map(point), do: { :ok, Geo.JSON.decode(point) }
-  def cast(point) when is_binary(point), do: { :ok, Geo.JSON.decode(point) }
+  def cast(point) when is_binary(point), do: { :ok, Poison.decode!(point) |> Geo.JSON.decode }
   def cast(_), do: :error
 
 end